Split up fixtures
