+++ tests/fixtures/helpers.py
@@ -1,0 +1,31 @@
+"""Fixtures for testing helper functions."""
+
+from __future__ import annotations
+
+import typing
+
+import pytest
+
+from tests.fixtures.utils import delete_data
+
+if typing.TYPE_CHECKING:
+    import pathlib
+    from collections.abc import Generator
+
+    import xnat
+
+
+@pytest.fixture
+def remove_destination_test_data(
+    connection_destination: xnat.BaseXNATSession,
+    xnat_root_dirs: dict[str, pathlib.Path],
+) -> Generator[xnat.BaseXNATSession, None, None]:
+    """Fixture to delete data on destination and metadata dir e.g. output/localhost."""
+    yield
+    delete_data(connection_destination, xnat_root_dirs["destination"])
+
+
+@pytest.fixture
+def unique_username(request: pytest.FixtureRequest) -> str:
+    """Generate a unique username based on the test name."""
+    return request.node.name.replace("[", "_").replace("]", "_")
