+++ tests/fixtures/utils.py
@@ -1,0 +1,140 @@
+"""Helper functions for fixtures."""
+
+from __future__ import annotations
+
+import logging
+import pathlib
+import shutil
+import subprocess
+import time
+import typing
+
+import requests  # type: ignore[import-untyped]
+
+import xnat4tests
+
+if typing.TYPE_CHECKING:
+    import xnat
+
+logger = logging.getLogger(__name__)
+
+
+def delete_data(
+    session: xnat.XNATSession,
+    destination_xnat_root_dir: pathlib.Path,
+) -> None:
+    """
+    Fixture calls this function.
+
+    Deletes all subject data (projects can't be deleted if re-running tests with same container).
+    Also, deletes metadata folder.
+
+    Parameters
+    ----------
+    session
+        The xnatpy sessions.
+
+    """
+    for project in session.projects:
+        for subject in project.subjects.values():
+            if project.id != "OPENNEURO_T1W" or subject.label != "dummydicomsubject":
+                session.delete(
+                    path=f"/data/projects/{project.id}/subjects/{subject.label}",
+                    query={"removeFiles": "True"},
+                )
+        project.subjects.clearcache()
+
+    archive_path = destination_xnat_root_dir / "archive"
+
+    if archive_path.exists():
+        for project in archive_path.iterdir():
+            if project.is_dir():
+                shutil.rmtree(project)
+
+    metadata_folder = pathlib.Path(__file__).resolve().parents[1] / "src" / "xmigrate" / "output" / "localhost"
+    if metadata_folder.exists():
+        shutil.rmtree(str(metadata_folder))
+
+
+def install_plugin(
+    jar_path: pathlib.Path,
+    plugin_dir: pathlib.Path,
+    connection_name: str,
+    config: xnat4tests.Config,
+) -> None:
+    """Install plugin for specified connection."""
+    # Check existing plugins
+    result = subprocess.run(  # noqa: S603
+        ["docker", "exec", connection_name, "ls", plugin_dir.as_posix()],  # noqa: S607
+        check=True,
+        capture_output=True,
+        text=True,
+    )
+    plugins_list = result.stdout.splitlines()
+
+    # If already installed → do nothing
+    if jar_path.name in plugins_list:
+        return
+
+    # Otherwise copy plugin
+    try:
+        subprocess.run(  # noqa: S603
+            [  # noqa: S607
+                "docker",
+                "cp",
+                str(jar_path),
+                f"{connection_name}:{(plugin_dir / jar_path.name).as_posix()}",
+            ],
+            check=True,
+        )
+    except subprocess.CalledProcessError as e:
+        msg = f"Command {e.cmd} failed with {e.returncode}: {e.output}"
+        raise RuntimeError(msg) from e
+
+    # Only restart if we actually installed something
+    xnat4tests.restart_xnat(config)
+
+
+def wait_for_connection(config: xnat4tests.Config) -> xnat.BaseXNATSession:
+    """
+    Wait for XNAT to become available and datatypes to be ready.
+
+    Tries up to max_retries times:
+      - Connects if not already connected
+      - Checks datatypes
+    """
+    max_retries = 30
+    conn = None
+
+    for attempt in range(1, max_retries + 1):
+        try:
+            # Connect if needed
+            if conn is None:
+                conn = xnat4tests.connect(config)
+                msg = f"[{attempt}] Connected to XNAT"
+                logger.info(msg)
+
+            # Check if datatypes are ready
+            datatypes = conn.get("/xapi/schemas/datatypes").json()
+            if len(datatypes) > 0:
+                msg = f"[{attempt}] Datatypes ready"
+                logger.info(msg)
+                return conn  # success
+            msg = f"[{attempt}] Datatypes not ready yet"
+            logger.info(msg)
+
+        except (requests.ReadTimeout, requests.ConnectionError) as e:
+            msg = f"[{attempt}] Connection failed, will retry: {e}"
+            logger.info(msg)
+            conn = None  # force reconnect next loop
+        except (requests.RequestException, ValueError) as e:
+            msg = f"[{attempt}] Datatypes request failed, retrying: {e}"
+            logger.info(msg)
+
+        time.sleep(1)
+
+    # Clean up if never ready
+    if conn is not None:
+        conn.disconnect()
+    msg = f"XNAT never became ready after {max_retries} attempts"
+    raise RuntimeError(msg)
